Move get_cmd_squares from UnitOrders to UnitManeuver
